refactor(cli): move status command to app/cli/status.py
